requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.1.0
numpy>=1.26.0
pyyaml>=6.0.1
jinja2>=3.1.2
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
# Optional:
# sendgrid>=6.11.0
# boto3>=1.34.0
# psycopg2-binary>=2.9.9
